Move ListController to components/controller
