+++ dashboard/src/components/partials/EventsList/EventsList.js
@@ -15,7 +15,7 @@ import TableList, {
 } from "/components/TableList";
 
 import Loader from "/components/util/Loader";
-import ListController from "/components/util/ListController";
+import ListController from "/components/controller/ListController";
 
 import Pagination from "/components/partials/Pagination";
 import SilenceEntryDialog from "/components/partials/SilenceEntryDialog";
